requirements.txt
pandas
mlflow
numpy
notebook
scikit-learn
matplotlib
python-box
pyYAML
tqdm
joblib
types-pyYAML
Flask
Flask-Cors
python-dotenv